removed unique constraint on trip id in vehicle position feed
This means that if duplicate trip ids are encountered in a csv input, the entire feed will not be invalidated. Circumstances this could occur is with non timetabled unknown trips (U001 etc) which could possibly appear in multiple places in the network, or if trip ids are changed and multiple instances exist for some operational reason.